projects/marketing_agency/dev_products/development/wishket_auto_bid/requirements.txt
anthropic>=0.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
